feat: Parse context of calls used for API usage
Currently, the `improper_api_usage` smell detector often returns false positives because it is matching only on the final function or method name making the call.

However, many APIs share common names (e.g., `get`, `append`, etc), so it’s easy for this to trigger even if the method is bound to a different module / class.

While the correct way to solve this would be to get the types involved in each call and then resolve the call context using the type, this isn’t available with the current AST. So instead, this commit uses the information in the current AST to build out the context for the call (i.e., the symbol(s) the call is being made against, if any).

Right now this commit retain the old behavior of just matching on the name, but in the next commit, I’ll add a toggle so the full context for API calls can be used to detect improper API usage.