feat: Add a switch to enable full API call paths
When using just the call name, a lot of false positives can occur, since many APIs share common method names.

To resolve this, the last commit added the ability to detect the context of each call, and this commit adds a switch that uses that context when detecting multiple API calls.

This isn’t perfect, and may create a few false negatives. Since we don’t have type information, it’s not possible to accurately detect the API a call belongs to. Specifically, assigning the same API to different variable names can cause repeated calls to this API to not be detected.

Such false negatives should be pretty minimal, however, since people tend to use the same variable names for the same API within the same file.