feat: Allow excluding API call paths from API usage
There are certain calls that are common to use in all code, whose presence doesn’t necessarily indicate a problem. For instance, `get` is commonly used to retrieve values from a variety of data structures. This provides a way to exclude these calls from being considered for the Potential Improper API Usage smell.