feat: Add existing call ratio into smell message
If a user wants to tune the detector for Potential Improper API Usage, they need to know what repetitive call ratio is being used to compare against the repetition threshold.

Note that `repetitive_calls` will never be `None` in this code, so the new code skips that check and simply calculates the ratio.